requirements.txt
attrs==19.1.0
certifi==2019.3.9
chardet==3.0.4
Click==7.0
dash==0.40.0
dash-core-components==0.45.0
dash-html-components==0.15.0
dash-renderer==0.21.0
dash-table==3.6.0
decorator==4.4.0
Flask==1.0.2
Flask-Compress==1.4.0
happybase==1.1.0
idna==2.8
ipython-genutils==0.2.0
itsdangerous==1.1.0
Jinja2==2.10
jsonschema==3.0.1
jupyter-core==4.4.0
kafka-python==1.4.5
MarkupSafe==1.1.1
nbformat==4.4.0
nltk==3.4
numpy==1.16.2
oauthlib==3.0.1
pandas==0.24.2
plotly==3.7.1
ply==3.11
py4j==0.10.7
pyrsistent==0.14.11
PySocks==1.6.8
pyspark==2.4.0
python-dateutil==2.8.0
pytz==2018.9
requests==2.21.0
requests-oauthlib==1.2.0
retrying==1.3.3
singledispatch==3.4.0.3
six==1.12.0
thriftpy==0.3.9
tqdm==4.31.1
traitlets==4.3.2
tweepy==3.7.0
urllib3==1.24.1
Werkzeug==0.15.1
